// management/frontend/src/api/schemas/userCreate.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface UserCreate {
  /**
     * @minLength 1
     * @maxLength 150
     */
  username: string;
  avatar?: string | null;
  status?: number;
  is_active?: boolean;
  is_root?: boolean;
  last_login?: string | null;
  /**
     * @minLength 5
     * @maxLength 128
     */
  password: string;
}