// management/frontend/src/api/zod/system-users/system-users.ts
export const systemUsersCreateUserBodyUsernameMax = 150;

export const systemUsersCreateUserBodyAvatarOneMax = 255;

export const systemUsersCreateUserBodyStatusDefault = 1;
export const systemUsersCreateUserBodyIsActiveDefault = true;
export const systemUsersCreateUserBodyIsRootDefault = false;
export const systemUsersCreateUserBodyPasswordMin = 5;
export const systemUsersCreateUserBodyPasswordMax = 128;



export const SystemUsersCreateUserBody = zod.object({
  "username": zod.string().min(1).max(systemUsersCreateUserBodyUsernameMax),
  "avatar": zod.union([zod.string().max(systemUsersCreateUserBodyAvatarOneMax),zod.null()]).optional(),
  "status": zod.number().default(systemUsersCreateUserBodyStatusDefault),
  "is_active": zod.boolean().default(systemUsersCreateUserBodyIsActiveDefault),
  "is_root": zod.boolean().default(systemUsersCreateUserBodyIsRootDefault),
  "last_login": zod.union([zod.iso.datetime({}),zod.null()]).optional(),
  "password": zod.string().min(systemUsersCreateUserBodyPasswordMin).max(systemUsersCreateUserBodyPasswordMax)
})
// ...
export const systemUsersUpdateUserBodyUsernameMax = 150;

export const systemUsersUpdateUserBodyAvatarOneMax = 255;

export const systemUsersUpdateUserBodyStatusDefault = 1;
export const systemUsersUpdateUserBodyIsActiveDefault = true;
export const systemUsersUpdateUserBodyIsRootDefault = false;
export const systemUsersUpdateUserBodyPasswordOneMin = 5;
export const systemUsersUpdateUserBodyPasswordOneMax = 128;



export const SystemUsersUpdateUserBody = zod.object({
  "username": zod.string().min(1).max(systemUsersUpdateUserBodyUsernameMax),
  "avatar": zod.union([zod.string().max(systemUsersUpdateUserBodyAvatarOneMax),zod.null()]).optional(),
  "status": zod.number().default(systemUsersUpdateUserBodyStatusDefault),
  "is_active": zod.boolean().default(systemUsersUpdateUserBodyIsActiveDefault),
  "is_root": zod.boolean().default(systemUsersUpdateUserBodyIsRootDefault),
  "last_login": zod.union([zod.iso.datetime({}),zod.null()]).optional(),
  "password": zod.union([zod.string().min(systemUsersUpdateUserBodyPasswordOneMin).max(systemUsersUpdateUserBodyPasswordOneMax),zod.null()]).optional()
})